requirements.txt
fastapi~=0.68.1
tortoise-orm~=0.17.7
xlwt~=1.3.0
uvicorn~=0.15.0
python-dotenv~=0.19.0